+++ rasgl/src/core/include/rasgl/core/graphics.h
@@ -32,6 +32,6 @@ typedef struct RenderState {
 /**
  * Project a point in view space to screen space.
  */
-Point2i project_point(int32_t screen_width, int32_t screen_height, int32_t projection_ratio, Point3f view_point);
+Point2f project_point(int32_t screen_width, int32_t screen_height, int32_t projection_ratio, Point3f view_point);
 
 #endif
